Fix Specialization when a side is provided
Specialization function accept a side as argument that indicates the
supposed side of the original module. This can not always be provided
and is used only as an optimisation, to avoid copying module
signatures.

Care must be taken with provided path, since those might
already have been specialized. Only sides inside the environment
should be trusted.